% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stab_mixed_parallel.R
\name{stab_mixed_parallel}
\alias{stab_mixed_parallel}
\title{Mixed model stability parallel computation}
\usage{
stab_mixed_parallel(
  data,
  combinations,
  cores,
  progress,
  response,
  target,
  fixed,
  random,
  p.threshold,
  icc.interval,
  icc.type,
  n.sims
)
}
\arguments{
\item{data, }{a tidy data set provided in the wrapper function}

\item{combinations, }{a set of combination as created by the wrapper function}

\item{cores, }{number of cores to be used in parallel execution}

\item{progress, }{should a progress bar be shown}

\item{response, }{the response variable (e.g. log(expression))}

\item{target, }{character vector for target column in data}

\item{fixed, }{formula for the fixed effects (not including target)}

\item{random, }{formula in lme4-style for the random effect (only one accepted)}

\item{p.threshold, }{threshold for effect of Beta (design variables)}

\item{icc.interval, }{specifying interval for the ICC CI}

\item{icc.type, }{type of ICC passed to boot function}

\item{n.sims, }{number of simulations in the ICC boot estimation}
}
\value{
A data frame with cominations of genes and their ICC confidence interval.
}
\description{
This function returns a data frame of ICC calculated from gene sets of size n. To be used with the
exported wrapper function.
}
